+++ meier/blog/presentation/view/assets.py
