+++ meier/static/js/admin/user/user.js
@@ -14,7 +14,7 @@ var vm = new Vue({
             url:'/admin/user/api/user_info',
             responseType:'application/json'
         }).then(function (res) {
-            var payload = res.data.data;
+            let payload = res.data.data;
             vm.email = payload.email;
             vm.user_name = payload.user_name;
             vm.twitter_profile= payload.twitter_profile;
@@ -22,15 +22,13 @@ var vm = new Vue({
             vm.website= payload.website;
             vm.user_desc= payload.user_desc;
         }).catch(function (err) {
-            // todo : error
             console.log(err);
         });
     },
     methods: {
         update: function (event) {
             console.log("update");
-
-            var params = {
+            let params = {
                 email:vm.email,
                 user_name:vm.user_name,
                 twitter_profile:vm.twitter_profile,
