+++ meier/static/js/blog/random_post.js
@@ -1,0 +1,13 @@
+function moveRandomPostURL(){
+    let headers = {};
+    Requests.get(
+        apiURL.randomPostAPI,
+        headers,
+        moveRandomPostURLSuccessCallBack,
+        function (){}
+    );
+}
+
+function moveRandomPostURLSuccessCallBack(response) {
+    location.href = response.data.data.url;
+}
