+++ meier/templates/themes/zen_tree/head.html
@@ -27,5 +27,10 @@
 <link rel="stylesheet"
       href="//cdn.jsdelivr.net/gh/highlightjs/cdn-release@11.6.0/build/styles/default.min.css">
 <script src="//cdn.jsdelivr.net/gh/highlightjs/cdn-release@11.6.0/build/highlight.min.js"></script>
+<script src="//cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 <link rel="stylesheet" type="text/css" href="/assets/css/github-dark.css"/>
 <script>hljs.highlightAll();</script>
+
+<!-- meier -->
+<script src="/js/blog/base.js" type="text/javascript"></script>
+<script src="/js/blog/random_post.js" type="text/javascript"></script>
