+++ meier/templates/themes/zen_tree/post_detail.html
@@ -36,6 +36,15 @@
             crossorigin="anonymous"
             async>
     </script>
+    <script>
+        document.addEventListener('keydown', (event) => {
+            // key
+            if (event.ctrlKey && event.key === 'r') {
+                moveRandomPostURL();
+                return false;
+            }
+        });
+    </script>
 </div>
 </body>
 </html>
